% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{browser}
\alias{browser}
\alias{values}
\alias{labels}
\alias{groups}
\alias{qColors}
\title{Browser market share example data}
\usage{
data("browser", package = "rhtmlDonut")
}
\description{
Example data used to demonstrate \code{\link{Donut}}. Describes the market
share of web browsers, broken down by browser version.
}
\details{
Loading this dataset attaches four objects:
\describe{
  \item{\code{values}}{numeric vector of length 187. Market share of each
    browser version, in percent.}
  \item{\code{labels}}{character vector of length 187. Browser name and
    version, e.g. \code{"Chrome 43.0"}.}
  \item{\code{groups}}{character vector of length 187. Browser family of
    each entry, e.g. \code{"Chrome"}.}
  \item{\code{qColors}}{character vector of length 14. Hex color codes
    suitable for the \code{values.color} argument.}
}
}
\keyword{datasets}
