% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhtmlDonut.R
\name{Donut}
\alias{Donut}
\title{Create a Donut plot}
\usage{
Donut(values, values.font = NULL, values.size = 10, values.color = NULL,
  values.display = "percentage", value.thres = NULL, labels,
  labels.font = NULL, labels.size = 10, labels.color = NULL,
  labels.minFontSize = 8, labels.inner = FALSE, groups = NULL,
  groups.font = NULL, groups.size = 10, groups.color = NULL,
  prefix = NULL, suffix = NULL, order = "default", order.control = TRUE,
  border.color = "white", max.label.length = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{values.display}{choice of c("percentage", "original"). If "percentage" then values are converted to percentages. If "original" display the original data. Default is "percentage".}

\item{value.thres}{threshold of the minimum value in percentage that will have a label attached. Range is [0,100] and default is 0.3}

\item{order}{ordering of the plot = c("default", "initial", "alphabetical", "descending")}

\item{border.color}{c("white", "none", hex colors)}
}
\value{
a donut plot
}
\description{
Create a Donut plot
}

